tests/test_main.py
import unittest
from main import recognize_speech, process_command, respond

class TestMainFunctions(unittest.TestCase):

    def test_recognize_speech(self):
        # This test will be a placeholder as we cannot test speech recognition without actual audio input
        self.assertTrue(True)

    def test_process_command(self):
        command = "Turn on the lights"
        expected_output = ['Turn', 'lights']
        self.assertEqual(process_command(command), expected_output)

    def test_respond(self):
        # This test will be a placeholder as we cannot test text-to-speech without actual audio output
        self.assertTrue(True)

if __name__ == '__main__':
    unittest.main()
